refactor project structure and add make targets
