+++ Makefile
@@ -6,3 +6,11 @@ sample:
 # Build the binary and put it in a bin dir
 build:
 	go build -o bin/ .
+
+# compile go code to wasm binary
+wasm:
+	GOOS=js GOARCH=wasm go build -o ./assets/jwt.wasm ./cmd/wasm/.
+
+# run go server
+run:
+	go run cmd/server/main.go
