+++ cmd/server/main.go
@@ -6,7 +6,7 @@ import (
 )
 
 func main() {
-	err := http.ListenAndServe(":8080", http.FileServer(http.Dir("../assets")))
+	err := http.ListenAndServe(":8080", http.FileServer(http.Dir("./assets")))
 	if err != nil {
 		fmt.Println("Failed to start server", err)
 		return
