+++ pkg/jwt/decode.go
@@ -1,4 +1,4 @@
-package lib
+package jwt
 
 import (
 	"encoding/base64"
